// src/app/core/api/model/imageResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ImageResponseDto { 
    id: string;
    appId: string;
    imageRef: string;
    commitSha: string;
    branch: string;
    githubPackageId?: string;
    sizeBytes?: number;
    fluiTags: Array<string>;
    isCurrentlyDeployed: boolean;
    createdAt: string;
}